pyproject.toml
[project]
name = "pytrans"
version = "0.1.0"
description = ""
authors = [
    {name = "Mukwano",email = "sparenica@afrowave.ltd"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "textual-dev (>=1.7.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "pytrans", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
